style(db): move prisma client to its own file
